% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sovi_data_kab_514_15}
\alias{sovi_data_kab_514_15}
\alias{sovi_data_kab_514_15.xlsx}
\title{Data SoVI 514 Kabupaten/Kota Indonesia (2015)}
\format{
Data frame dengan 514 baris dan 17 kolom:
\describe{
  \item{DISTRICTCODE}{Kode unik wilayah Kabupaten/Kota (ID wilayah).}
  \item{KABUPATEN}{Nama Kabupaten/Kota.}
  \item{AGE014}{Persentase penduduk usia 0--14 tahun (\%).}
  \item{FEMPOP}{Persentase penduduk perempuan (\%).}
  \item{AGE65P}{Persentase penduduk usia 65 tahun ke atas (\%).}
  \item{FEMHH}{Persentase rumah tangga dengan kepala keluarga perempuan (\%).}
  \item{HHSIZE}{Rata-rata jumlah anggota rumah tangga (jiwa).}
  \item{NOELEC}{Persentase rumah tangga tanpa akses listrik (\%).}
  \item{LOWEDUC}{Persentase penduduk dengan pendidikan rendah (\%).}
  \item{POPGRW}{Tingkat pertumbuhan penduduk (\%).}
  \item{POOR}{Persentase penduduk miskin (\%).}
  \item{ILLIT}{Tingkat buta huruf (\%).}
  \item{NOTRAIN}{Persentase penduduk tanpa pelatihan kerja formal (\%).}
  \item{DISAREA}{Persentase luas wilayah rawan bencana (\%).}
  \item{RENTHH}{Persentase rumah tangga yang menyewa tempat tinggal (\%).}
  \item{NOSAN}{Persentase rumah tangga tanpa fasilitas sanitasi layak (\%).}
  \item{NOWATER}{Persentase rumah tangga tanpa akses air bersih (\%).}
}
}
\source{
Data diolah dari Badan Pusat Statistik (BPS) Indonesia, tahun 2015.
Metodologi mengacu pada Cutter et al. (2003).
}
\description{
Dataset Social Vulnerability Index (SoVI) untuk 514 Kabupaten/Kota
di Indonesia tahun 2015, berisi 15 indikator sosial-ekonomi dan kerentanan.
}
\examples{
# Akses dataset sampel
path <- system.file("extdata", "sovi_data_kab_514_15.xlsx",
                    package = "soviclust")
df <- readxl::read_excel(path)
head(df[, 1:5])

}
\references{
Cutter, S.L., Boruff, B.J., & Shirley, W.L. (2003).
Social Vulnerability to Environmental Hazards.
\emph{Social Science Quarterly}, 84(2), 242--261.
\doi{10.1111/1540-6237.8402002}
}
