% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sovi_data_pop_514}
\alias{sovi_data_pop_514}
\alias{sovi_data_pop_514.xlsx}
\title{Data Populasi 514 Kabupaten/Kota Indonesia}
\format{
Data frame dengan 514 baris dan minimal 2 kolom:
\describe{
  \item{DISTRICTCODE}{Kode unik wilayah Kabupaten/Kota.}
  \item{POPULATION}{Jumlah penduduk (jiwa). Semua nilai positif.}
}
}
\description{
Dataset jumlah penduduk untuk 514 Kabupaten/Kota di Indonesia.
Digunakan sebagai vektor bobot populasi pada analisis FGWC, LFGWC,
dan ALFGWC.
}
\examples{
# Akses data populasi sampel
path <- system.file("extdata", "sovi_data_pop_514.xlsx",
                    package = "soviclust")
pop_df <- readxl::read_excel(path)
# Ekstrak vektor populasi (kolom numerik pertama)
pop <- as.numeric(pop_df[[which(sapply(pop_df, is.numeric))[1]]])
summary(pop)

}
