% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\name{sovi_sample_pop}
\alias{sovi_sample_pop}
\title{Baca Data Populasi 514 Kabupaten/Kota}
\usage{
sovi_sample_pop()
}
\value{
`data.frame` dengan 514 baris dan minimal 2 kolom (ID + populasi).
}
\description{
Membaca data jumlah penduduk 514 Kabupaten/Kota Indonesia
yang disertakan dalam package. Data ini digunakan sebagai bobot
populasi pada analisis FGWC, LFGWC, dan ALFGWC.
}
\examples{
pop_df <- sovi_sample_pop()
head(pop_df)

# Ambil vektor populasi
pop <- pop_df[[which(sapply(pop_df, is.numeric))[1]]]
summary(pop)
hist(pop / 1e6, main = "Distribusi Populasi (juta jiwa)",
     xlab = "Populasi (juta)", col = "steelblue")

}
