% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\name{sovi_sample_shapefile}
\alias{sovi_sample_shapefile}
\title{Baca Shapefile 514 Kabupaten/Kota Indonesia}
\usage{
sovi_sample_shapefile()
}
\value{
Objek `sf` dengan 514 fitur (Polygon/MultiPolygon).
  Kolom ID wilayah: `idkab`.
  Sistem koordinat: WGS84 (EPSG:4326).
}
\description{
Membaca shapefile batas wilayah 514 Kabupaten/Kota Indonesia
yang disertakan dalam package sebagai objek `sf`.
}
\examples{
shp <- sovi_sample_shapefile()
class(shp)
names(shp)
nrow(shp)

# Tampilkan peta sederhana
plot(sf::st_geometry(shp))

# Gabungkan dengan data SoVI
df  <- sovi_sample_data()
shp_sovi <- merge(shp, df, by.x = "idkab", by.y = "DISTRICTCODE")

}
