% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sovi_data_kab_514_15}
\alias{sovi_data_kab_514_15}
\alias{sovi_data_kab_514_15.xlsx}
\title{SoVI Data for 514 Indonesian Districts (2015)}
\format{
A data frame with 514 rows and 17 columns:
\describe{
  \item{DISTRICTCODE}{Unique district ID code.}
  \item{KABUPATEN}{District name.}
  \item{AGE014}{Percentage of population aged 0--14 years (\%).}
  \item{FEMPOP}{Percentage of female population (\%).}
  \item{AGE65P}{Percentage of population aged 65 years and above (\%).}
  \item{FEMHH}{Percentage of households with female head of household (\%).}
  \item{HHSIZE}{Average household size (persons).}
  \item{NOELEC}{Percentage of households without electricity access (\%).}
  \item{LOWEDUC}{Percentage of population with low education level (\%).}
  \item{POPGRW}{Population growth rate (\%).}
  \item{POOR}{Percentage of population living in poverty (\%).}
  \item{ILLIT}{Illiteracy rate (\%).}
  \item{NOTRAIN}{Percentage of population without formal job training (\%).}
  \item{DISAREA}{Percentage of district area classified as disaster-prone (\%).}
  \item{RENTHH}{Percentage of households renting their dwelling (\%).}
  \item{NOSAN}{Percentage of households without adequate sanitation (\%).}
  \item{NOWATER}{Percentage of households without clean water access (\%).}
}
}
\source{
Data compiled from Statistics Indonesia (BPS), 2015.
Methodology follows Cutter et al. (2003).
}
\description{
Social Vulnerability Index (SoVI) dataset for 514 Indonesian districts
(Kabupaten/Kota) in 2015, containing 15 socioeconomic and vulnerability indicators.
}
\examples{
# Access the sample dataset
path <- system.file("extdata", "sovi_data_kab_514_15.xlsx",
                    package = "soviclust")
df <- readxl::read_excel(path)
head(df[, 1:5])

}
\references{
Cutter, S.L., Boruff, B.J., & Shirley, W.L. (2003).
Social Vulnerability to Environmental Hazards.
\emph{Social Science Quarterly}, 84(2), 242--261.
\doi{10.1111/1540-6237.8402002}
}
