% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sovi_data_pop_514}
\alias{sovi_data_pop_514}
\alias{sovi_data_pop_514.xlsx}
\title{Population Data for 514 Indonesian Districts}
\format{
A data frame with 514 rows and at least 2 columns:
\describe{
  \item{DISTRICTCODE}{Unique district ID code.}
  \item{POPULATION}{Total population (persons). All values are positive.}
}
}
\description{
Population count data for 514 Indonesian districts. Used as a population
weight vector in FGWC, LFGWC, and ALFGWC clustering analyses.
}
\examples{
# Access sample population data
path <- system.file("extdata", "sovi_data_pop_514.xlsx",
                    package = "soviclust")
pop_df <- readxl::read_excel(path)
# Extract population vector (first numeric column)
pop <- as.numeric(pop_df[[which(sapply(pop_df, is.numeric))[1]]])
summary(pop)

}
