% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\name{sovi_sample_coords}
\alias{sovi_sample_coords}
\title{Read Centroid Coordinates for 514 Indonesian Districts}
\usage{
sovi_sample_coords()
}
\value{
A `data.frame` with 514 rows.
}
\description{
Reads bundled centroid longitude and latitude data.
}
\examples{
coord <- sovi_sample_coords()
head(coord)
dim(coord)

}
