% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\name{sovi_sample_distmat}
\alias{sovi_sample_distmat}
\title{Read Distance Matrix for 514 Indonesian Districts}
\usage{
sovi_sample_distmat()
}
\value{
A numeric 514 x 514 matrix.
}
\description{
Reads the bundled inter-district distance matrix.
}
\examples{
mat <- sovi_sample_distmat()
dim(mat)
diag(mat)[1:5]

}
