% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\name{sovi_sample_shapefile}
\alias{sovi_sample_shapefile}
\title{Read Administrative Boundaries for 514 Indonesian Districts}
\usage{
sovi_sample_shapefile()
}
\value{
An `sf` object with 514 spatial features.
}
\description{
Reads the bundled administrative-boundary object for 514 Indonesian
regencies/cities. The spatial object is stored internally as an RDS file
to reduce package size.
}
\examples{
shp <- sovi_sample_shapefile()
class(shp)
nrow(shp)
names(shp)

}
